populated initQuestions function
populated initQuestions function. It will now assign values to the variables when it creates an object based on automated system that uses values that make sense for this test procedure. The actual question data would need to be added manually or pulled from a JSON file but this will check if the system works correctly so far.

Will test once initAnswers function is finished as this needs to be called first to populate the array needed for the initQuestions function.