Added pick question function
Added pick question function. This picks random questions from the question array and adds them to the array to use for the set amount of questions. This function will randomise the order the same number of questions are drawn in so that it will be hard to display the exact same set of questions time after time.

This could be optimised with two seperate arrays, one for assigned questions and one that have yet to be assigned as this would reduced the amount of searching needed on the boolean array. May look into this at a later point.